wip(objects): more reorganisation
