+++ Objects/Converters/ConverterRevit/ConverterRevitShared/Partial Classes/ConvertBrace.cs
@@ -1,5 +1,5 @@
 ﻿using Autodesk.Revit.DB.Structure;
-using Objects.BuildingObject.PhysicalObjects;
+using Objects.BuildingObject;
 using Speckle.Core.Models;
 using System.Collections.Generic;
 using DB = Autodesk.Revit.DB;
