+++ Objects/Objects/BuildingObject/Rebar.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 using Objects.BuildingObject.enums;
+using Objects.Definitions;
 
-namespace Objects.BuildingObject.PhysicalObjects
+namespace Objects.BuildingObject
 {
   public class Rebar :CurveBasedElement
   {
