+++ Objects/Objects/BuildingObject/Room.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Objects.ProjectOrganization;
+using Objects.Definitions;
+using Objects.Organization;
 
-namespace Objects.BuildingObject.PhysicalObjects
+namespace Objects.BuildingObject
 {
   public class Room : PointBasedElement
   {
