+++ Objects/Objects/BuildingObject/Space.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Objects.Definitions;
 
-namespace Objects.BuildingObject.PhysicalObjects
+namespace Objects.BuildingObject
 {
-  public class Station : PointBasedElement
+  public class Space : PointBasedElement
   {
     // to implement source app parameters interface from claire
   }
