--- Objects/Objects/BuildingObject/enums/AxisEnum.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Objects.BuildingObject.enums
-{
-  public enum AxisType
-  {
-    Cartesian,
-    Cylindrical,
-    Spherical
-  }
-}
