+++ Objects/Objects/Definitions/BaseCurveProperty.cs
@@ -4,13 +4,13 @@ using System.Text;
 using Objects.BuildingObject.enums;
 using Speckle.Core.Models;
 
-namespace Objects.BuildingObject.Calculations
+namespace Objects.Definitions
 {
-  public class PointProperty : Base
+  public class BaseCurveProperty : Base
   {
     public string name { get; set; }
+    //public double thickness { get; set; }
 
-    public PointPropertyType type { get; set; }
-
+    public CurveElementType type { get; set; }
   }
 }
