+++ Objects/Objects/Definitions/PointBasedElement.cs
@@ -1,22 +1,26 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
-using Objects.BuildingObject.Calculations;
+using Objects.BuildingObject.enums;
 using Speckle.Core.Models;
 using Objects.Geometry;
 using Objects.Visualization;
 
-namespace Objects.BuildingObject.PhysicalObjects
+namespace Objects.Definitions
 {
-  public abstract class PointBasedElement : Base , IDisplayValue<List<Mesh>> 
+  public abstract class PointBasedElement : Base, IDisplayValue<List<Mesh>>
   {
     public Point basePoint { get; set; }
-    //public string name { get; set; }
+
     public PointProperty property { get; set; }
     public Material material { get; set; }
     public Base childrenObject { get; set; } //maybe holding pointers here would make sense more
     public List<Mesh> displayValue { get; set; }
   }
 
+  public class PointProperty : Base
+  {
+    public string name { get; set; }
 
+    public PointPropertyType type { get; set; }
+  }
 }
