+++ Objects/Objects/Structural/Geometry/Storey.cs
@@ -3,7 +3,7 @@ using Speckle.Core.Kits;
 using Speckle.Core.Models;
 using System.Collections.Generic;
 using Objects.Geometry;
-using Objects.ProjectOrganization;
+using Objects.Organization;
 
 namespace Objects.Structural.Geometry
 {
