+++ Objects/Objects/System/CableTray.cs
@@ -1,5 +1,6 @@
-﻿using Objects.BuildingObject.PhysicalObjects;
-using Objects.ProjectOrganization;
+﻿using Objects.BuildingObject;
+using Objects.Definitions;
+using Objects.Organization;
 
 namespace Objects.System
 {
