+++ Objects/Objects/Utils/Parameters.cs
@@ -1,5 +1,4 @@
-﻿using Objects.BuiltElements.Revit;
-using Speckle.Core.Models;
+﻿using Speckle.Core.Models;
 using System;
 using System.Collections.Generic;
 using System.Text;
